+++ R/add_synonyms.R
@@ -14,8 +14,9 @@ add_synonyms <- function(gene_map,
                          syn_map) {
 
     ## Avoid confusing Biocheck
-    input_gene <- NULL
-
+    input_gene <- NULL;
+    
+    if(is.null(syn_map)) return(gene_map)
     gene_map <- dplyr::rename(gene_map, input_gene_standard = input_gene)
     syn_dict <- setNames(syn_map$input, syn_map$name)
     gene_map <- dplyr::mutate(gene_map,