+++ NEWS.md
@@ -6,6 +6,7 @@
     - Can now take a user-supplied gene_map to use instead of gprofiler/homologene/babelgene.
     - Lets user take advantage of all of the many:many ortholog mapping functions for organisms
     not covered in any of the databases (eg. earthworm).
+* New helper function: `set_gprofiler` (to use beta version)
 
 ## Bug fixes